+++ packages/server/controllers/apps.js
@@ -1,5 +1,5 @@
 import { logger } from '@appsemble/node-utils';
-import { normalize, StyleValidationError, validateStyle } from '@appsemble/utils';
+import { normalize, permissions, StyleValidationError, validateStyle } from '@appsemble/utils';
 import Boom from '@hapi/boom';
 import Ajv from 'ajv';
 import crypto from 'crypto';
@@ -9,6 +9,7 @@ import { col, fn, literal, Op, UniqueConstraintError } from 'sequelize';
 import sharp from 'sharp';
 import * as webpush from 'web-push';
 
+import checkRole from '../utils/checkRole';
 import getAppBlocks from '../utils/getAppBlocks';
 import getAppFromRecord from '../utils/getAppFromRecord';
 import getDefaultIcon from '../utils/getDefaultIcon';
@@ -138,6 +139,7 @@ export async function createApp(ctx) {
       throw Boom.forbidden('User does not belong in this organization.');
     }
 
+    await checkRole(ctx, OrganizationId, permissions.CreateApps);
     await checkBlocks(definition, db);
 
     for (let i = 1; i < 11; i += 1) {
@@ -282,6 +284,7 @@ export async function updateApp(ctx) {
       throw Boom.forbidden("User does not belong in this App's organization.");
     }
 
+    await checkRole(ctx, OrganizationId, [permissions.EditApps, permissions.EditAppSettings]);
     await dbApp.update(result, { where: { id: appId } });
 
     ctx.body = getAppFromRecord({ ...dbApp.dataValues, ...result });
@@ -376,6 +379,23 @@ export async function patchApp(ctx) {
       throw Boom.forbidden("User does not belong in this App's organization.");
     }
 
+    const checkPermissions = [];
+
+    if (
+      domain !== undefined ||
+      path !== undefined ||
+      isPrivate !== undefined ||
+      template !== undefined
+    ) {
+      checkPermissions.push(permissions.EditAppSettings);
+    }
+
+    if (yaml || definition) {
+      checkPermissions.push(permissions.EditApps);
+    }
+
+    await checkRole(ctx, dbApp.OrganizationId, checkPermissions);
+
     await dbApp.update(result, { where: { id: appId } });
 
     ctx.body = getAppFromRecord({ ...dbApp.dataValues, ...result });
@@ -400,6 +420,8 @@ export async function deleteApp(ctx) {
   if (!organizations.some(organization => organization.id === app.OrganizationId)) {
     throw Boom.forbidden("User does not belong in this App's organization.");
   }
+
+  await checkRole(ctx, app.OrganizationId, permissions.DeleteApps);
 
   await app.update({ path: null });
   await app.destroy();
@@ -485,6 +507,8 @@ export async function setAppIcon(ctx) {
     throw Boom.forbidden("User does not belong in this App's organization.");
   }
 
+  await checkRole(ctx, app.OrganizationId, permissions.EditAppSettings);
+
   await app.update({ icon });
   ctx.status = 204;
 }
@@ -505,6 +529,8 @@ export async function deleteAppIcon(ctx) {
   if (!organizations.some(organization => organization.id === app.OrganizationId)) {
     throw Boom.forbidden("User does not belong in this App's organization.");
   }
+
+  await checkRole(ctx, app.OrganizationId, permissions.EditAppSettings);
 
   await app.update({ icon: null });
 
@@ -548,6 +574,8 @@ export async function broadcast(ctx) {
   if (!user.organizations.some(organization => organization.id === app.OrganizationId)) {
     throw Boom.forbidden('User does not belong in this app’s organization.');
   }
+
+  await checkRole(ctx, app.OrganizationId, permissions.PushNotifications);
 
   const { vapidPublicKey: publicKey, vapidPrivateKey: privateKey } = app;
 
@@ -657,6 +685,8 @@ export async function setAppBlockStyle(ctx) {
     if (!block) {
       throw Boom.notFound('Block not found.');
     }
+
+    await checkRole(ctx, app.OrganizationId, permissions.EditApps);
 
     if (css.length) {
       await AppBlockStyle.upsert({